requirements.txt
Django==4.2.16
djangorestframework==3.15.2
psycopg2-binary==2.9.9
pendulum==3.0.0
drf-yasg==1.21.7
gunicorn==22.0.0